rename: 캐러셀 버튼 폴더명 수정
