+++ src/pages/Main/components/CarouselArrow/index.tsx
